Refactor Paginator
